moved validation function to helpers
