+++ clustertools/shared/helpers.py
@@ -1,11 +1,13 @@
 from __future__ import annotations
+
+import time
 from functools import wraps
 from os.path import expanduser, expandvars, realpath
 from pathlib import Path
 from typing import overload, TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from clustertools.shared.typing import PathLike
+    from clustertools.shared.typing import PathLike, WallTimeStr
 
 
 @overload
@@ -26,3 +28,17 @@ def bindable(func):
     def bind(instance):
         return func.__get__(instance)
     return bind
+
+
+def validate_walltime(walltime_str: str) -> WallTimeStr:
+    try:
+        time.strptime(walltime_str, '%H:%M:%S')
+    except ValueError:
+        try:
+            time.strptime(walltime_str, '%M:%S')
+        except ValueError:
+            raise ValueError(
+                "Received malformed 'wall_time' string. Format should be "
+                "'%H:%M:%S', or '%M:%S' if requesting < 1 hour"
+            )
+    return walltime_str
