+++ clustertools/shared/typing.py
@@ -1,4 +1,3 @@
-import time
 from pathlib import Path, PurePosixPath
 import sys
 from typing import (BinaryIO,
@@ -26,19 +25,6 @@ MswStderrDest = Union[MswDestBase, Type[type(sys.stderr)], Literal['stderr']]
 
 
 WallTimeStr = NewType('WallTimeStr', str)
-
-def validate_walltime(walltime_str: str) -> WallTimeStr:
-    try:
-        time.strptime(walltime_str, '%H:%M:%S')
-    except ValueError:
-        try:
-            time.strptime(walltime_str, '%M:%S')
-        except ValueError:
-            raise ValueError(
-                "Received malformed 'wall_time' string. Format should be "
-                "'%H:%M:%S', or '%M:%S' if requesting < 1 hour"
-            )
-    return WallTimeStr(walltime_str)
 
 
 # EmailAddress = NewType('EmailAddress', str)
